% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{annot.dat}
\alias{annot.dat}
\title{Annotates data}
\usage{
annot.dat(df, t)
}
\arguments{
\item{t}{}

\item{dat}{dataframe}
}
\value{
the annotated dataframe
}
\description{
This function performs annotation on a data frame by sorting unique indexes and then filtering the dataframe based on X, Y, and Z coordinates. The function then checks for certain conditions and assigns new indexes to rows that meet those conditions.
}
