% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensity_modeling.R
\name{betamix.2d}
\alias{betamix.2d}
\title{Calculate threshold from 2D array using beta mixture model}
\usage{
betamix.2d(arr, mixnum = 2)
}
\arguments{
\item{arr}{A 2D array of values}

\item{mixnum}{The number of mixture components to use in the beta mixture model. Default is 2.}
}
\value{
A threshold value
}
\description{
This function takes a 2D array and the number of mixture components as inputs, and returns a threshold value. It first extracts values above a certain intensity from the array, scales these values between 0.0001 and 0.9999, fits a beta mixture model to these scaled values, and calculates a density difference between the two beta distributions. Finally, it scales the intersection point of these distributions back to the original scale of the array values and returns it as the threshold.
}
\examples{
arr <- matrix(c(0, 0, 1, 5, 10, 20), nrow = 2, ncol = 3)
betamix.2d(arr)
}
