% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensity_modeling.R
\name{betamix.3d}
\alias{betamix.3d}
\title{Calculate threshold from 3D array using beta mixture model}
\usage{
betamix.3d(arr, mixnum = 2, n.cores = 1, ...)
}
\arguments{
\item{arr}{A 3D array of values}

\item{mixnum}{The number of mixture components to use in the beta mixture model. Default is 2.}

\item{n.cores}{number of cores used in parallel computing. Default is 1.}

\item{...}{additional parameters passed to \code{mclapply}}
}
\value{
A threshold value
}
\description{
This function takes a 3D array, the number of mixture components, number of cores and additional parameters as inputs, and returns a threshold value. It applies \code{betamix.2d} function on each slice of 3D array and  returns the threshold value of each slice.
}
\examples{
arr <- array(c(0, 0, 1, 5, 10, 20), c(2,3,3))
betamix.3d(arr)
}
