% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensity_modeling.R
\name{betamix_4d}
\alias{betamix_4d}
\title{Calculate threshold from 4D array using beta mixture model}
\usage{
betamix_4d(arr, mixnum = 2, n.cores = 1, ...)
}
\arguments{
\item{arr}{A 4D array of values}

\item{mixnum}{The number of mixture components to use in the beta mixture model. Default is 2.}

\item{n.cores}{number of cores used in parallel computing. Default is 1.}

\item{...}{additional parameters passed to \code{mclapply}}
}
\value{
A threshold value
}
\description{
This function takes a 4D array, the number of mixture components, number of cores and additional parameters as inputs, and returns a threshold value. It applies \code{betamix_3d} function on each time point of 4D array and  returns the threshold value of each time point.
}
