% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensity_modeling.R
\name{betamix_img}
\alias{betamix_img}
\title{Calculate threshold from image using beta mixture model}
\usage{
betamix_img(arr, mixnum = 2, n.cores = 1, ...)
}
\arguments{
\item{arr}{image array}

\item{mixnum}{The number of mixture components to use in the beta mixture model. Default is 2.}

\item{n.cores}{number of cores used in parallel computing. Default is 1.}

\item{...}{additional parameters passed to \code{betamix_3d()} and \code{betamix_4d()} functions}
}
\value{
A threshold value
}
\description{
This function takes an image array, number of mixture components, number of cores and additional parameters as inputs, and returns a threshold value. It detects the dimension of the image array and applies \code{betamix_2d()}, \code{betamix_3d()}, or \code{betamix_4d()} accordingly to calculate the threshold.
}
