% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boss_betamix.R
\name{boss_betamix}
\alias{boss_betamix}
\title{Obtain intersection value to binarize the image into background and ROI}
\usage{
boss_betamix(img, min.intensity, mix.num, threshold.percentile = 0.8)
}
\arguments{
\item{img}{Image to be binarized}

\item{min.intensity}{Minimum intensity value to be considered to be ROI}

\item{mix.num}{Number of the clusters to be considered in mixture model.}

\item{threshold.percentile}{Thresholding value if the mixture model is failed to be recognized.}
}
\value{
Thresholding value
}
\description{
Obtain intersection value to binarize the image into background and ROI
}
\examples{
boss_betamix(img, 0.8, 2)

}
