% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{connect.components}
\alias{connect.components}
\title{Connected components in an array}
\usage{
connect.components(roi_mask, r = 21)
}
\arguments{
\item{roi_mask}{array to find connected components in}

\item{r}{radius of the connected component filter}
}
\value{
labeled array of the same size
}
\description{
The function finds the connected components in an array and returns a labeled array of the same size
}
\examples{
connect.components(arr, 10)
}
