% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{get.delta.n}
\alias{get.delta.n}
\title{Get number of connected components in each timepoint of 4D array}
\usage{
get.delta.n(cell_df)
}
\arguments{
\item{cell_df}{dataframe of tracked components}
}
\value{
number of connected components in each timepoint
}
\description{
The function returns the number of connected components in each timepoint of a 4D array
}
\examples{
get.delta.n(cell_df)
}
