% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{get_adjacent}
\alias{get_adjacent}
\title{Get adjacent values}
\usage{
get_adjacent(arr, k)
}
\arguments{
\item{arr}{array to subset}

\item{k}{minimum number of adjacent values}
}
\value{
unique values that span at least k cells over last dimension of the array
}
\description{
The function returns values that span at least k cells over last dimension of the array
}
\examples{
get_adjacent(arr, 2)
returns values that span at least k cells over `d`, i.e. last dim. of arr
}
