% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{get_centroid}
\alias{get_centroid}
\title{Get centroid of a value in an array}
\usage{
get_centroid(arr, label)
}
\arguments{
\item{arr}{array}

\item{label}{value to get centroid of}
}
\value{
centroid of elements in array that match label
}
\description{
The function returns the centroid of elements in an array that match a specific value
}
\examples{
get_centroid(arr, 10)
}
