% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{get_k_long_labels}
\alias{get_k_long_labels}
\title{Gets k-long values}
\usage{
get_k_long_labels(arr, k)
}
\arguments{
\item{arr}{array to subset}

\item{k}{minimum number of adjacent values}
}
\value{
unique values that span at least k cells over last dimension of the array
}
\description{
The function returns values that span at least k cells over last dimension of the array
}
\examples{
get_k_long_labels(arr, 2)
}
