% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{get_size}
\alias{get_size}
\title{Get size of a value in an array}
\usage{
get_size(arr, label)
}
\arguments{
\item{arr}{array}

\item{label}{value to get size of}
}
\value{
sum of elements in array that match label
}
\description{
The function returns the sum of elements in an array that match a specific value
}
\examples{
get_size(arr, 10)
}
