% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rboss.R
\name{getbeta}
\alias{getbeta}
\title{Get regression coefficient}
\usage{
getbeta(data, point, type = 2)
}
\arguments{
\item{data}{A data frame which contains 3 columns for cox regression : survival time, censor status, Biomarker values.
2 columns for linear regression :  Y, X.}

\item{point}{A number to cut biomarker or X.}

\item{type}{A number; if =1, will perform linear regression; if =2(default) will perform cox regression.}
}
\value{
An object with 3 class: Coefficient beta, number of samples of which the biomarker is greater than the point, standard error of coefficient estimation.
}
\description{
Computes the regression coefficient of certain regression based on certain cutoff.
}
