% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rboss.R
\name{getpvalue}
\alias{getpvalue}
\title{Computes the distribution function of the multivariate normal distribution}
\usage{
getpvalue(threshold, mu, n, Sigma)
}
\arguments{
\item{threshold}{A number.}

\item{mu}{The mean vector of length n.}

\item{n}{A number indicates dimension.}

\item{Sigma}{The correlation matrix of dimension n.}
}
\value{
The evaluated distribution function
}
\description{
Computes the distribution function of the multivariate normal distribution.
}
