% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{last_ind}
\alias{last_ind}
\title{Subset based on last dimension}
\usage{
last_ind(arr, n)
}
\arguments{
\item{arr}{array to subset}

\item{n}{index of last dimension to subset on}
}
\value{
array subsetted on last dimension at index \code{n}
}
\description{
The function returns the values of an array at a specific index along the last dimension of the array
}
\examples{
last_ind(arr, 5)
}
