% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{list2nifti}
\alias{list2nifti}
\title{Convert a list of arrays to a Nifti object}
\usage{
list2nifti(list, dims)
}
\arguments{
\item{list}{A list of arrays that will be converted to a Nifti object.}

\item{dims}{A vector of dimensions for the resulting array.}
}
\value{
A Nifti object created from the input list of arrays.
}
\description{
list2nifti() converts a list of arrays to a Nifti object using the oro.nifti package.
}
\examples{
list2nifti(list(array(1:12, c(3, 4)), array(13:24, c(3, 4))), c(3, 4, 2))

# convert a list of 3D arrays to a 4D Nifti object
list2nifti(list(array(1:12, c(3, 4, 3)), array(13:24, c(3, 4, 3))), c(3, 4, 3, 2))
}
