% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{make.overlay}
\alias{make.overlay}
\title{Make Overlay}
\usage{
make.overlay(cell_df, dims, t)
}
\arguments{
\item{cell_df}{A dataframe containing the cell data, with columns for x, y, z, t, and index}

\item{dims}{A numeric vector of length 4, representing the dimensions of the image in x, y, z, and t}

\item{t}{An integer representing the timepoint of interest}
}
\value{
A 3-dimensional binary array representing the overlay
}
\description{
This function creates an overlay by drawing boxes around centroids of cells in a dataframe, at a specified timepoint, as well as cells in adjacent timepoints.
}
\examples{
\dontrun{
make.overlay(cell_df, dims, t)
}

}
