% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{make_cell_df}
\alias{make_cell_df}
\title{Create dataframe for a value in an array}
\usage{
make_cell_df(arr, label)
}
\arguments{
\item{arr}{array}

\item{label}{value to make dataframe of}
}
\value{
dataframe of size and centroid of elements in array that match label
}
\description{
The function creates a dataframe from the values of size and centroid of elements in an array that match a specific value
}
