% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_filtering.R
\name{median.filtering.4d}
\alias{median.filtering.4d}
\title{Median filter a 4D image}
\usage{
\method{median}{filtering.4d}(roi_mask, r = 10, n.cores = 1, ...)
}
\arguments{
\item{roi_mask}{4D array of image}

\item{r}{kernel size for median filter}

\item{n.cores}{number of cores to use for parallel processing}
}
\value{
4D array of median filtered image
}
\description{
The function applies a median filter to each time point of the 4D image using median.filtering.3d function.
}
\examples{
median.filtering.4d(roi_mask, 5, 2)
}
