% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_filtering.R
\name{median.filtering}
\alias{median.filtering}
\title{Median filter an image}
\usage{
\method{median}{filtering}(roi_mask, r = 10, n.cores = 1, ...)
}
\arguments{
\item{roi_mask}{image to be median filtered (3D or 4D array)}

\item{r}{kernel size for median filter}

\item{n.cores}{number of cores}
}
\description{
The function takes an image and applies the appropriate median filter function
(median.filtering.3d or median.filtering.4d) based on the dimensionality of the image.
}
