% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_filtering.R
\name{median_filtering_3d}
\alias{median_filtering_3d}
\title{Median filter a 3D image}
\usage{
median_filtering_3d(roi_mask, r = 10)
}
\arguments{
\item{roi_mask}{3D array of image}

\item{r}{kernel size for median filter}
}
\value{
3D array of median filtered image
}
\description{
The function applies a median filter to the 3D image using mmand::medianFilter.
}
