\name{myGene}
\docType{data}
\alias{myGene}
\title{clinical dataset}
\description{
  This data set gives the expression levels of gene data, overall survival time, and censoring
status from 500 lung adenocarcinoma cases.
}
\usage{myGene}
\format{A dataframe containing 500 observations of 3 variables.}
\source{ raw survival data come from https://tau.cmmt.ubc.ca/cSurvival/project_data/TCGA-LUAD/df_survival_o.csv and raw gene expression data come from  https://tau.cmmt.ubc.ca/cSurvival/project_data/TCGA-LUAD/df_gene.csv}
\references{
  Xuanjin Cheng, Yongxing Liu, Jiahe Wang, Yujie Chen, Andrew Gordon Robertson, Xuekui Zhang, Steven J M
Jones, and Stefan Taubert. (2022) \emph{cSurvival: a web resource for biomarker interactions in cancer outcomes and in cell lines}. Briefings in Bioinformatics.
}
\keyword{datasets}
