% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{plot_overlay}
\alias{plot_overlay}
\title{Plots a 3D image with an overlay}
\usage{
plot_overlay(img, overlay, z = 1)
}
\arguments{
\item{img}{3D nifti or array, usually the input of the boss model}

\item{overlay}{an overlay that is the result of \code{make_overlay}}

\item{z}{an integer vector representing the slices to be plotted}
}
\value{
a plot where the cell bodies are marked by the a red overlay
}
\description{
\code{plot_overlay} is a wrapper to \code{oro.nifti::overlay}. It takes a 3D image and an overlay that is the output of \code{make_overlay()} and plots the chosen slices.
}
