% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{post.process.df}
\alias{post.process.df}
\title{Postprocceses a data.frame that is the result of track.components}
\usage{
post.process.df(df, t)
}
\arguments{
\item{df}{dataframe}

\item{t}{}
}
\value{
the annotated dataframe
}
\description{
This function performs annotation on a data frame by sorting unique indexes and then filtering the dataframe based on X, Y, and Z coordinates. The function then checks for certain conditions and assigns new indexes to rows that meet those conditions.
}
