% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{post_process_df}
\alias{post_process_df}
\title{Postprocceses a data.frame that is the result of track_components}
\usage{
post_process_df(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
the annotated dataframe
}
\description{
This function performs annotation on a data frame by sorting unique indexes and then filtering the dataframe based on X, Y, and Z coordinates. The function then checks for certain conditions and assigns new indexes to rows that meet those conditions.
}
