% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholding.R
\name{threshold.img}
\alias{threshold.img}
\title{Threshold an image}
\usage{
threshold.img(img, thr, n.cores = 1, ...)
}
\arguments{
\item{img}{image to be thresholded (2D, 3D, or 4D array)}

\item{thr}{threshold value (scalar, vector, or matrix)}

\item{n.cores}{number of cores to use for parallel processing (only used in threshold.3d and threshold.4d)}
}
\value{
binary mask of thresholded image (same dimensionality as input image)
}
\description{
The function takes an image and a threshold value and applies the appropriate threshold function
(threshold.2d, threshold.3d, or threshold.4d) based on the dimensionality of the image.
It also checks that the threshold value has the correct dimensionality and length to match the image.
}
\examples{
threshold.img(img, 150)
threshold.img(img, c(120, 130, 140), 2)
threshold.img(img, matrix(c(120, 130, 140, 130, 140, 150), ncol = 2), 2)
}
