% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholding.R
\name{threshold_2d}
\alias{threshold_2d}
\title{Threshold a 2D image}
\usage{
threshold_2d(img, thr)
}
\arguments{
\item{img}{2D array of image}

\item{thr}{threshold value}
}
\value{
binary mask where pixels with values greater than the threshold are set to 1 and all others are set to 0
}
\description{
The function applies a threshold to the 2D image and creates a binary mask where all the pixels
with values greater than the threshold are set to 1 and the rest are set to 0.
}
