% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholding.R
\name{threshold_3d}
\alias{threshold_3d}
\title{Threshold a 3D image}
\usage{
threshold_3d(img, thr, n.cores = 1, ...)
}
\arguments{
\item{img}{3D array of image}

\item{thr}{vector of threshold values (one for each slice of the image)}

\item{n.cores}{number of cores to use for parallel processing}
}
\value{
3D binary mask
}
\description{
The function applies a threshold to each slice of the 3D image using threshold_2d function,
creating a 3D binary mask where all the pixels with values greater than the threshold in each slice are set to 1 and the rest are set to 0.
}
