% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholding.R
\name{threshold_4d}
\alias{threshold_4d}
\title{Threshold a 4D image}
\usage{
threshold_4d(img, thr, n.cores = 1, ...)
}
\arguments{
\item{img}{4D array of image}

\item{thr}{matrix of threshold values (one for each slice and time point of the image)}

\item{n.cores}{number of cores to use for parallel processing}
}
\value{
4D binary mask
}
\description{
The function applies a threshold to each slice and time point of the 4D image using threshold_2d function,
creating a 4D binary mask where all the pixels with values greater than the threshold in each slice and time point are set to 1 and the rest are set to 0.
}
