% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{track_components}
\alias{track_components}
\title{Track connected components in 4D array}
\usage{
track_components(roi_labels, k = 2, size.thr = 10)
}
\arguments{
\item{roi_labels}{labeled array of connected components}

\item{k}{minimum number of adjacent frames for a component to be tracked}

\item{size.thr}{minimum size of component to be tracked}
}
\value{
dataframe of tracked components
}
\description{
The function tracks the connected components in a 4D array and returns a dataframe of the tracked components
}
