% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotate_df}
\alias{annotate_df}
\title{Annotate data.frame}
\usage{
annotate_df(df, t)
}
\arguments{
\item{df}{A dataframe}

\item{t}{An integer value}
}
\value{
A dataframe with 4 columns: count, addition, deletion, and survivor.
}
\description{
This function takes in a dataframe that is result of post.process.df (or track.components)
and an integer t (total timepoints), and returns a dataframe with 4 columns:
count, addition, deletion, and survivor. The count column
contains the number of rows in the input dataframe where the "t" column has
a value equal to j, for each j in 1 to t. The addition column contains the
number of new rows in the dataframe between time j and j + 1. The deletion
column contains the number of rows in the dataframe at time j that do not
appear in the dataframe at any later time. The survivor column contains
the number of rows in the dataframe at time j that also appear in the dataframe
at time j + 1.
}
