% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotate_df}
\alias{annotate_df}
\title{Annotate data.frame}
\usage{
annotate_df(df, t)
}
\arguments{
\item{df}{A dataframe}

\item{t}{An integer value}
}
\value{
A dataframe with 4 columns: count, addition, deletion, and survivor.
}
\description{
This function takes in a dataframe that is result of post.process.df (or track.components)
and an integer t (total timepoints), and returns a dataframe with 4 columns:
\code{count}, \code{added}, \code{deleted}, and \code{survivor}. \code{count} is the number of cells at timepoint t.
\code{added} is the number of new cells in timepoint t (relative to timepoint t - 1);
\code{added} is always NA for t = 1.
\code{deleted} is the number of cells at t that do not show up at later timepoints.
\code{survivor} is the number of cells at timepoint t that remain from timepoint 1.
}
